+++ app/controllers/transfers/projects_controller.rb
@@ -24,9 +24,12 @@ class Transfers::ProjectsController < ApplicationController
       :outgoing_trust_ukprn,
       :establishment_sharepoint_link,
       :advisory_board_date,
+      :advisory_board_conditions,
       :provisional_transfer_date,
       :incoming_trust_sharepoint_link,
-      :outgoing_trust_sharepoint_link
+      :outgoing_trust_sharepoint_link,
+      :assigned_to_regional_caseworker_team,
+      :handover_note_body
     )
   end
 end