+++ app/forms/transfer/create_project_form.rb
@@ -25,6 +25,7 @@ class Transfer::CreateProjectForm < CreateProjectForm
       incoming_trust_sharepoint_link: incoming_trust_sharepoint_link,
       outgoing_trust_sharepoint_link: outgoing_trust_sharepoint_link,
       advisory_board_date: advisory_board_date,
+      advisory_board_conditions: advisory_board_conditions,
       transfer_date: provisional_transfer_date,
       regional_delivery_officer_id: user.id,
       team: user.team,
@@ -34,11 +35,14 @@ class Transfer::CreateProjectForm < CreateProjectForm
       tasks_data: Transfer::TasksData.new
     )
 
-    if valid?
-      @project.save!
-      return @project
+    return nil unless valid?
+
+    ActiveRecord::Base.transaction do
+      @project.save
+      @note = Note.create(body: handover_note_body, project: @project, user: user, task_identifier: :handover) if handover_note_body
     end
-    nil
+
+    @project
   end
 
   def provisional_transfer_date=(hash)